+++ packages/terra-clinical-header/tests/jest/Header.test.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import Header from '../../src/Header';
 
+const consoleSpy = jest.spyOn(global.console, 'warn');
+
 it('should render a default component', () => {
   const header = render(<Header />);
   expect(header).toMatchSnapshot();
@@ -34,7 +36,9 @@ it('should render a header with all content', () => {
 });
 
 it('should render a subheader with title and heading level', () => {
-  const subheader = render(<Header level={1} text="title" isSubheader />);
+  const subheader = render(<Header level={1} title="title" isSubheader />);
+  const titleWarningMessage = 'The `title` prop has been renamed to `text`. Please update all references of `title` prop to `text`.';
+  expect(consoleSpy).toHaveBeenCalledWith(titleWarningMessage);
   expect(subheader).toMatchSnapshot();
 });
 
@@ -51,7 +55,9 @@ it('should render a subheader with all content', () => {
   expect(subheader).toMatchSnapshot();
 });
 
-it('should render a header with no text from level not set', () => {
-  const header = render(<Header text="This title should not render without a provided level" />);
+it('should render a header with default heading level when level not set', () => {
+  const header = render(<Header text="This title should render with a default level" />);
+  const levelWarningMessage = 'Default heading level may not appropriate has it would fail to convey context of heading in a site / application where it is used. Heading level should be set explicitly depending on the position of header in site / application to allow screen readers to identify headers consistently.';
+  expect(consoleSpy).toHaveBeenCalledWith(levelWarningMessage);
   expect(header).toMatchSnapshot();
 });